requirements.txt
streamlit
deepface
opencv-python==4.11.0.86
Pillow
numpy==1.23.5
scipy
python-dotenv
supabase
requests
tensorflow==2.12.0
protobuf
ultralytics

